+++ defitrack-rest/defitrack-protocol-services/defitrack-polycat/src/main/java/io/defitrack/protocol/polycat/staking/PolycatStakingMarketService.kt
@@ -92,7 +92,7 @@ class PolycatStakingMarketService(
                 PriceRequest(
                     stakedtoken.address,
                     getNetwork(),
-                    balance.asEth(stakedtoken.decimals).toBigDecimal(),
+                    balance.asEth(stakedtoken.decimals),
                     stakedtoken.type
                 )
             )
